requirements.txt
agate==1.12.0
agate-dbf==0.2.3
agate-excel==0.4.1
agate-sql==0.7.2
babel==2.16.0
blinker==1.9.0
branca==0.8.1
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.8
csvkit==2.0.1
dbfread==2.0.7
et-xmlfile==1.1.0
Flask==3.1.0
folium==0.19.5
geographiclib==2.0
geopy==2.4.1
gunicorn==23.0.0
idna==3.10
isodate==0.6.1
itsdangerous==2.2.0
Jinja2==3.1.6
leather==0.4.0
MarkupSafe==3.0.2
numpy==2.1.1
olefile==0.47
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
parsedatetime==2.6
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
pytimeparse==1.1.8
pytz==2024.2
requests==2.32.3
six==1.16.0
SQLAlchemy==2.0.35
text-unidecode==1.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
Werkzeug==3.1.3
xlrd==2.0.1
xyzservices==2025.1.0
